Move movie list info into Home.js
