+++ src/App.js
@@ -1,43 +1,5 @@
-import { useEffect, useState } from "react";
-import Movie from "./components/Movie.js";
-
 function App() {
-  const [loading, setLoading] = useState(true);
-  const [movies, setMovies] = useState([]);
-
-  const getMovies = async () => {
-    const res = await (
-      await fetch(
-        "https://yts.mx/api/v2/list_movies.json?minimum_rating=8.8&sort_by=year"
-      )
-    ).json();
-
-    setMovies(res.data.movies);
-    setLoading(false);
-  };
-  useEffect(() => {
-    getMovies();
-  }, []);
-
-  return (
-    <div>
-      {loading ? (
-        <h3>Loading...</h3>
-      ) : (
-        <div className="wrapper__movies">
-          {movies.map((movie) => (
-            <Movie
-              key={movie.id}
-              coverImg={movie.medium_cover_image}
-              title={movie.title}
-              summary={movie.summary}
-              genres={movie.genres}
-            />
-          ))}
-        </div>
-      )}
-    </div>
-  );
+  return null;
 }
 
 export default App;
